# frontend/scripts/openapi/contracts/auth.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebShellSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "hostId": {
            "format": "uuid",
            "title": "Hostid",
            "type": "string"
          },
          "hostName": {
            "title": "Hostname",
            "type": "string"
          },
          "idleTimeoutSeconds": {
            "title": "Idletimeoutseconds",
            "type": "integer"
          },
          "maxDurationSeconds": {
            "title": "Maxdurationseconds",
            "type": "integer"
          },
          "sshPort": {
            "title": "Sshport",
            "type": "integer"
          },
          "streamPath": {
            "title": "Streampath",
            "type": "string"
          },
          "ticketExpiresAt": {
            "format": "date-time",
            "title": "Ticketexpiresat",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          },
          "webShellSessionId": {
            "format": "uuid",
            "title": "Webshellsessionid",
            "type": "string"
          }
        },
        "required": [
          "webShellSessionId",
          "hostId",
          "hostName",
          "address",
          "sshPort",
          "username",
          "streamPath",
          "ticketExpiresAt",
          "idleTimeoutSeconds",
          "maxDurationSeconds"
        ],
        "title": "WebShellSessionResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "OpsComposer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/hosts/{host_id}/web-shell-sessions": {
      "post": {
        "operationId": "createWebShellSession",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Host Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebShellSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Web Shell Session",
        "tags": [
          "web-shell"
        ]
      }
    },
...
    "/api/v1/web-shell-sessions/{web_shell_session_id}": {
      "delete": {
        "operationId": "closeWebShellSession",
        "parameters": [
          {
            "in": "path",
            "name": "web_shell_session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Web Shell Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close Web Shell Session",
        "tags": [
          "web-shell"
        ]
      }
    },
...
  }
}